NF: _burySiblings in Abstract
