+++ AnkiDroid/src/main/java/com/ichi2/libanki/sched/AbstractSched.java
@@ -3,6 +3,7 @@ package com.ichi2.libanki.sched;
 import android.app.Activity;
 import android.content.Context;
 import android.content.res.Resources;
+import android.database.Cursor;
 import android.util.Pair;
 import android.widget.Toast;
 
@@ -14,6 +15,8 @@ import com.ichi2.libanki.Consts;
 import com.ichi2.libanki.Deck;
 import com.ichi2.libanki.DeckConfig;
 import com.ichi2.libanki.Collection;
+import com.ichi2.libanki.Utils;
+import com.ichi2.libanki.utils.Time;
 import com.ichi2.utils.JSONArray;
 import com.ichi2.utils.JSONObject;
 
@@ -434,6 +437,7 @@ public abstract class AbstractSched {
      */
     public abstract void unsuspendCards(long[] ids);
     public abstract void buryCards(long[] cids);
+
     @VisibleForTesting
     public abstract void buryCards(long[] cids, boolean manual);
     /**
@@ -622,4 +626,52 @@ public abstract class AbstractSched {
     }
 
     protected abstract JSONObject _newConf(Card card);
+
+    /**
+     * Sibling spacing
+     * ********************
+     */
+
+    protected void _burySiblings(Card card) {
+        ArrayList<Long> toBury = new ArrayList<>();
+        JSONObject nconf = _newConf(card);
+        boolean buryNew = nconf.optBoolean("bury", true);
+        JSONObject rconf = _revConf(card);
+        boolean buryRev = rconf.optBoolean("bury", true);
+        // loop through and remove from queues
+        Cursor cur = null;
+        try {
+            cur = mCol.getDb().getDatabase().query(
+                    "select id, queue from cards where nid=? and id!=? "+
+                            "and (queue=" + Consts.QUEUE_TYPE_NEW + " or (queue=" + Consts.QUEUE_TYPE_REV + " and due<=?))",
+                    new Object[] {card.getNid(), card.getId(), mToday});
+            while (cur.moveToNext()) {
+                long cid = cur.getLong(0);
+                int queue = cur.getInt(1);
+                List<Long> queue_object;
+                if (queue == Consts.QUEUE_TYPE_REV) {
+                    queue_object = mRevQueue;
+                    if (buryRev) {
+                        toBury.add(cid);
+                    }
+                } else {
+                    queue_object = mNewQueue;
+                    if (buryNew) {
+                        toBury.add(cid);
+                    }
+                }
+                // even if burying disabled, we still discard to give
+                // same-day spacing
+                queue_object.remove(cid);
+            }
+        } finally {
+            if (cur != null && !cur.isClosed()) {
+                cur.close();
+            }
+        }
+        // then bury
+        if (!toBury.isEmpty()) {
+            buryCards(Utils.collection2Array(toBury),false);
+        }
+    }
 }
